chore: move `deepmd.pt.infer.deep_eval.eval_model` to tests
Per discussion in https://github.com/deepmodeling/deepmd-kit/pull/4142#issuecomment-2359848991. It should not be a public API as it lacks maintainance.
