+++ source/tests/pt/model/test_forward_lower.py
@@ -4,9 +4,6 @@ import unittest
 
 import torch
 
-from deepmd.pt.infer.deep_eval import (
-    eval_model,
-)
 from deepmd.pt.model.model import (
     get_model,
 )
@@ -19,6 +16,9 @@ from deepmd.pt.utils.nlist import (
 
 from ...seed import (
     GLOBAL_SEED,
+)
+from ..common import (
+    eval_model,
 )
 from .test_permutation import (  # model_dpau,
     model_dpa1,
