+++ source/tests/pt/model/test_permutation.py
@@ -5,9 +5,6 @@ import unittest
 
 import torch
 
-from deepmd.pt.infer.deep_eval import (
-    eval_model,
-)
 from deepmd.pt.model.model import (
     get_model,
 )
@@ -17,6 +14,9 @@ from deepmd.pt.utils import (
 
 from ...seed import (
     GLOBAL_SEED,
+)
+from ..common import (
+    eval_model,
 )
 
 CUR_DIR = os.path.dirname(__file__)
