+++ source/tests/pt/model/test_rot_denoise.py
@@ -4,9 +4,6 @@ import unittest
 
 import torch
 
-from deepmd.pt.infer.deep_eval import (
-    eval_model,
-)
 from deepmd.pt.model.model import (
     get_model,
 )
@@ -16,6 +13,9 @@ from deepmd.pt.utils import (
 
 from ...seed import (
     GLOBAL_SEED,
+)
+from ..common import (
+    eval_model,
 )
 from .test_permutation_denoise import (
     model_dpa1,
